fix conversation name, check id in update - delete

+++ src/apps/controllers/products.controller.ts
@@ -5,9 +5,9 @@ class ProductsController{
     async getProductById(req: Request, res: Response){
         try{
             const {id} = req.params;
-            const success = await productsService.getProductById(Number(id));
-            if(success){
-                res.status(200).json(success);
+            const result = await productsService.getProductById(Number(id));
+            if(result){
+                res.status(200).json(result);
             }else{
                 res.status(404).json({message: "Product not found"});
             }
@@ -18,8 +18,8 @@ class ProductsController{
 
     async getAllProducts(req: Request, res: Response){
         try{
-            const success = await productsService.getAllProducts();
-            res.status(200).json(success);
+            const result = await productsService.getAllProducts();
+            res.status(200).json(result);
         }catch(err){
             res.status(500).json({ message: "Internal Server Error", error: err });
         }
@@ -28,9 +28,9 @@ class ProductsController{
     async createProduct(req: Request, res: Response){
         try{
             const newProduct = req.body;
-            const success = await productsService.createProduct(newProduct);
-            if(success){
-                res.status(201).json({message: "Product created successfully", product: success});
+            const result = await productsService.createProduct(newProduct);
+            if(result){
+                res.status(201).json({message: "Product created successfully", product: result});
             }else{
                 res.status(400).json({message: "Failed to create product"});
             }
@@ -39,12 +39,17 @@ class ProductsController{
         }
     }
 
-    async updateProduct(req: Request, res: Response){
+    async updateProduct(req: Request, res: Response): Promise<void>{
         try{
             const {id} = req.params;
+            const productId = await productsService.getProductById(Number(id));
+            if (!productId) {
+                res.status(404).json({ message: "Product not found" });
+                return 
+            }
             const updatedFields = req.body;
-            const success = await productsService.updateProduct(Number(id), updatedFields);
-            if(success){
+            const result = await productsService.updateProduct(Number(id), updatedFields);
+            if(result){
                 res.status(200).json({message: "Product updated successfully"});
             }else{
                 res.status(400).json({message: "Failed to update product"});
@@ -54,11 +59,16 @@ class ProductsController{
         }
     }
 
-    async deleteProduct(req: Request, res: Response){
+    async deleteProduct(req: Request, res: Response): Promise<void>{
         try{
             const {id} = req.params;
-            const success = await productsService.deleteProduct(Number(id));
-            if(success){
+            const productId = await productsService.getProductById(Number(id));
+            if (!productId) {
+                res.status(404).json({ message: "Product not found" });
+                return 
+            }
+            const result = await productsService.deleteProduct(Number(id));
+            if(result){
                 res.status(200).json({message: "Product deleted successfully"});
             }else{
                 res.status(404).json({message: "Product not found"});
@@ -71,8 +81,8 @@ class ProductsController{
     async searchProductsByKeyword(req: Request, res: Response){
         try{
             const { keyword } = req.query;
-            const success = await productsService.searchProducts(keyword as string);
-            res.status(200).json(success);
+            const result = await productsService.searchProducts(keyword as string);
+            res.status(200).json(result);
         }catch(err){
             res.status(500).json({ message: "Internal Server Error", error: err });
         }
